% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanLM.R
\name{CleanLM}
\alias{CleanLM}
\title{Clean Results from a Linear Regression Model}
\usage{
CleanLM(res)
}
\arguments{
\item{res}{A linear regression object.}
}
\value{
A table with two columns: the label of each variable level and the result \verb{B (95\%CI; p-value)}. Variable levels that do not appear in this table are the reference group.
}
\description{
This function cleans the results of any linear regression object, and returns them in format \verb{B (95\%CI; p-value)}.
}
\examples{
### Crude regression model
m_crude <- lm(y ~ var1, data = dat)
CleanLog(m_crude)
# rn                 logres
# 1 var1 1.12 (0.85 - 1.06; 0.097)

### Multivariable regression model
m_multi <- lm(y ~ var1 + var2 + var3, data = dat)
CleanLM(m_multi)
# rn                    logres
# 1  var1  1.12 (0.85 - 1.06; 0.097)
# 2  var2  -0.96 (-0.61 - 1.5; 0.849)
# 3  var3  0.99 (0.97 - 1.02; 0.639)
}
