% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanLog.R
\name{CleanLog}
\alias{CleanLog}
\title{\code{CleanLog()}}
\usage{
CleanLog(res)
}
\arguments{
\item{x}{A logistic regression object.}
}
\value{
A table with two columns: the label of each variable level and the result (95\%CI; p-value). Variable levels that do not appear in this table are the reference group.
}
\description{
This function cleans the results of any logistic regression object, and returns them in format OR (95\%CI; p-value).
}
\examples{
### Crude regression model
m_crude <- glm(outcome ~ var1, data = dat, family = "binomial")
CleanLog(m_crude)
# rn                 logres
# 1 var1 1.12 (0.85 - 1.06; 0.097)

### Multivariable regression model
m_crude <- glm(outcome ~ var1 + var2 + var3, data = dat, family = "binomial")
CleanLog(m_crude)
# rn                    logres
# 1  var1  1.12 (0.85 - 1.06; 0.097)
# 2  var2  0.96 (0.61 - 1.5; 0.849)
# 3  var3  0.99 (0.97 - 1.02; 0.639)
}
