% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqOutcome.R
\name{FreqOutcome}
\alias{FreqOutcome}
\title{Get number of events by group}
\usage{
FreqOutcome(gr_var, status_var, df, exposed_level = NULL)
}
\arguments{
\item{gr_var}{Single character vector with the name of the grouping variable in the data (e.g. treatment). It can be any class, but needs exactly 2 levels.}

\item{status_var}{Single character vector with the name of the outcome variable in the data (e.g. mortality). It has to be logical (i.e., TRUE / FALSE).}

\item{df}{Either a \code{data.frame} or a \code{tibble} containing both variables.}

\item{exposed_level}{Name of the exposed group in \code{gr_var} (e.g., with the treatment). Only required if the \code{gr_var} is a character vector or a factor.}
}
\value{
A small data.frame with the number of events by number of patients on each group (events / N (\%)).
}
\description{
A short description...
}
\examples{
### Data
test <- data.frame(
  id = c(1:10),
  death = c(F, T, T, F, T, T, F, F, T, F),
  tt = c("a", "b", "b", "a", "a", "a", "b", "b", "b", "a")
)

### Example
FreqOutcome(gr_var = "tt", status_var = "death", df = test, exposed_level = "a")
#               events / N (\%)
# 1     exposed    3 / 5 (60\%)
# 2 not exposed    2 / 5 (40\%)
}
